Re-organised notifiers into a separate package. Renamed 'oversight.py' as 'oversight_runner.py', to avoid conflict with the package name. Added support for a notification sink that stores the triggering image in a directory structure.
