+++ oversight/notifiers/image_file_notifier.py
@@ -1,0 +1,44 @@
+# Copyright 2017 Ben Carson. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+__author__ = 'bcarson'
+
+import logging
+import os
+
+from oversight.signals import trigger_event
+
+logger = logging.getLogger('root')
+
+class ImageFileNotifier(object):
+    """
+    A notification sink that responds to trigger events by storing the image file.
+    """
+    def __init__(self, labelled_images_root):
+        self.labelled_images_root = labelled_images_root
+
+        trigger_event.connect(self.handle_trigger_event)
+
+    def handle_trigger_event(self, sender, **data):
+        # Write the image into a file in under the image root.
+        filename = '%s-%s.jpg' % (data['source'], data['timestamp'].strftime('%Y%m%d-%H%M%S'))
+        image_path = '/'.join([self.labelled_images_root, data['prediction'], filename])
+
+        try:
+            file_handle = open(image_path, 'wb')
+            file_handle.write(data['image'])
+            file_handle.close()
+        except IOError, e:
+            logger.error("IO Error: ", e)
+
